+++ instrumentation/opentelemetry-instrumentation-flask/src/opentelemetry/instrumentation/flask/__init__.py
@@ -266,6 +266,9 @@ from opentelemetry.instrumentation.propagators import (
 )
 from opentelemetry.instrumentation.utils import _start_internal_or_server_span
 from opentelemetry.metrics import get_meter
+from opentelemetry.semconv._incubating.attributes.http_attributes import (
+    HTTP_TARGET,
+)
 from opentelemetry.semconv.attributes.http_attributes import HTTP_ROUTE
 from opentelemetry.semconv.metrics import MetricInstruments
 from opentelemetry.semconv.metrics.http_metrics import (
@@ -399,6 +402,9 @@ def _rewrapped_app(
                 duration_attrs_old[HTTP_ROUTE] = wrapped_app_environ.get(
                     _ENVIRON_REQUEST_ROUTE_KEY
                 )
+                duration_attrs_old[HTTP_TARGET] = wrapped_app_environ.get(
+                    _ENVIRON_REQUEST_ROUTE_KEY
+                )
 
             duration_histogram_old.record(
                 max(round(duration_s * 1000), 0), duration_attrs_old
@@ -410,6 +416,9 @@ def _rewrapped_app(
 
             if wrapped_app_environ.get(_ENVIRON_REQUEST_ROUTE_KEY, None):
                 duration_attrs_new[HTTP_ROUTE] = wrapped_app_environ.get(
+                    _ENVIRON_REQUEST_ROUTE_KEY
+                )
+                duration_attrs_new[HTTP_TARGET] = wrapped_app_environ.get(
                     _ENVIRON_REQUEST_ROUTE_KEY
                 )