+++ instrumentation/opentelemetry-instrumentation-flask/tests/test_programmatic.py
@@ -87,6 +87,10 @@ def expected_attributes_new(override_attributes):
 
 _server_duration_attrs_old_copy = _server_duration_attrs_old.copy()
 _server_duration_attrs_old_copy.append("http.route")
+_server_duration_attrs_old_copy.append("http.target")
+
+_server_duration_attrs_new_copy = _server_duration_attrs_new.copy()
+_server_duration_attrs_new_copy.append("http.target")
 
 _expected_metric_names_old = [
     "http.server.active_requests",
@@ -102,7 +106,7 @@ _recommended_metrics_attrs_old = {
 }
 _recommended_metrics_attrs_new = {
     "http.server.active_requests": _server_active_requests_count_attrs_new,
-    "http.server.request.duration": _server_duration_attrs_new,
+    "http.server.request.duration": _server_duration_attrs_new_copy,
 }
 _server_active_requests_count_attrs_both = (
     _server_active_requests_count_attrs_old
@@ -113,7 +117,7 @@ _server_active_requests_count_attrs_both.extend(
 _recommended_metrics_attrs_both = {
     "http.server.active_requests": _server_active_requests_count_attrs_both,
     "http.server.duration": _server_duration_attrs_old_copy,
-    "http.server.request.duration": _server_duration_attrs_new,
+    "http.server.request.duration": _server_duration_attrs_new_copy,
 }
 
 
@@ -574,6 +578,7 @@ class TestProgrammatic(InstrumentationTest, WsgiTestBase):
         expected_duration_attributes = {
             "http.method": "GET",
             "http.route": "/hello/<int:helloid>",
+            "http.target": "/hello/<int:helloid>",
             "http.host": "localhost",
             "http.scheme": "http",
             "http.flavor": "1.1",
@@ -600,6 +605,7 @@ class TestProgrammatic(InstrumentationTest, WsgiTestBase):
             "http.request.method": "GET",
             "url.scheme": "http",
             "http.route": "/hello/<int:helloid>",
+            "http.target": "/hello/<int:helloid>",
             "network.protocol.version": "1.1",
             "http.response.status_code": 200,
         }